logging: Add prompt to LOG2_ALWAYS_RUNTIME option
Prompt was removed by 154ca8526 to ensure that it is not
disable when it is required. That removed possibility of
manually enabling that option my the user. Bringing prompt
back and adding check to code to fail compilation if
LOG2_ALWAYS_RUNTIME is not set but is required.

(cherry picked from commit 3cbfc505b357f80b6aca4cd2dc39c0f60a24c1b1)

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 3cbfc505b357f80b6aca4cd2dc39c0f60a24c1b1
Change-Id: I43c9d7b587523d564f41be17bb90e97416e2a8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3513072
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>