kconfig: fix dt_node_has_prop and add nodelabel functions
The Kconfig function "dt_node_has_prop" was using label as its
parameter, where other functions use either chosen or path.
The documentation says that the parameter is path, so this patch
makes the function as documentation says and as other functions
in the file.
The additional nodelabel functions were added as counterparts that
are using nodes labels instead of paths.

Signed-off-by: Michał Barnaś <mb@semihalf.com>
GitOrigin-RevId: a1ab8da86243d6ebcd5116a22ed3ddf4270a9260
Change-Id: I48a4e7e9d3cf1edeb6d9d4a5e8f2385c03e5b94a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3552803
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>