linker: Create sections from zephyr,memory-region nodes
Currently when a node has a 'zephyr,memory-region' compatible and a
'zephyr,memory-region' string property, a new memory region is created
in the linker script.

Having a memory region without a section to place variables in could be
not that useful. With this patch we extend the memory-region mechanism
to also create sections.

The user can then place variables in the sections as usual by using for
example the GCC attributes.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: bc72fb67b1c81670a087f766ef38c3517da10366
Change-Id: Iefc943e769bbfacd4b211ff5ff25df47fa8fa6cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3538987
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>