riscv: isr.S: dedicate a register to &current_cpu
Stop using &_kernel as this is not SMP friendly. Let's use s0 (after
preserving its content) to hold &current_cpu instead so it won't have
to be reloaded 	each time it is needed. This will be even more relevant
when SMP support is added.

Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
GitOrigin-RevId: 4f5374854e88d96253c9d55cad75f76987bbadf6
Change-Id: Ifc110efc2f1193f04349a47e3052ed99fb60fa1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3542113
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>