posix: Fix pthread_once has incorrect behavior
As described in
The behavior of pthread_once() is undefined if once_control has
automatic storage duration or is not initialized by PTHREAD_ONCE_INIT

However, in Zephyr, the implementation is incorrect. If the init value
is PTHREAD_ONCE_INIT, the program will never run the init_func.

https://pubs.opengroup.org/onlinepubs/7908799/xsh/pthread_once.html.
Signed-off-by: Jaxson Han <jaxson.han@arm.com>
GitOrigin-RevId: 0e1ff84fe7acf0420aa430c4ef5e544f92f8b710
Change-Id: I48fe95145de11e9f2ee6188eac3ee57f4b22e858
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3573994
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>