cbprintf: Fix incorrect alignment
In addition to x86-32, POSIX boards are also affected by the issue
where a `long double` has the size of 12.

Simplify the process by setting the alignment to 16 for xtensa (as
before). Then defaulting to using `long double` if the config is set.
The case of 12 byte `long double` is then handled by Z_POW2_CEIL when
available. Otherwise, a BUILD_ASSERT is used to verify that alignment
is valid.

Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: 2a80d67843feb9dfc84d75371706a70cd306547d
Change-Id: I5419ea730fc08d2c2fa936bb413bfff458e236b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3552869
Tested-by: CopyBot Service Account <copybot.service@gmail.com>