code_relocation: Move NOCOPY flag to avoid OS drive colon
COPY/NOCOPY flag was added to the end of the string with ':'
as separators. The python script then split the line on ':'
which breaks on Windows build because the string was

[MEM]:[FILE]:[COPY]

In windows you will have 'c:\' in the file path so the
line.split() in script will split on the 'c:'.

Move the COPY/NOCOPY to:

[MEM]:[COPY/NOCOPY]:[FILE]

Note that the comments at top of gen_relocate_app.py also
indicates this format.

Fixes #43950

Signed-off-by: David Leach <david.leach@nxp.com>
GitOrigin-RevId: c243da02ed5c3612e70e0f5a4702903a87c55c7d
Change-Id: I384ca1e4b711377bc582fb85fa98670b13b51ca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3545529
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>