Bluetooth: Controller: radio: Fix num of antennas returned by controller
The antenna number returned by controller if antenna switching is not
enabled was zero. That is not compliant with  BT 5.3. Core Spec Vol 4.
Part E section 7.8.87. In this situation returned value should be 1.

There is always single antenna available that is responsible for PDU
reception and transmission.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>
GitOrigin-RevId: 09cd296962d871526bfd8f84da3aa13dec10811f
Change-Id: Ia866b94f776aad88fe340094f969905182209b52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3552902
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>