kernel: add `k_can_yield` helper function
Implements a function that application and driver code can use to check
whether it is valid to yield (or block) in the current context. This
check is required for functions that can feasibly be run from multiple
contexts. The primary intended use case is power management transition
functions, which can be run by application code explicitly or
automatically in the idle thread by system PM.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: 1ef647f396d8c26475807b6424b3269455c41411
Change-Id: I0562e9c52fb7108c0fabe488365c063519fae246
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3632357
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>