Bluetooth: CSIS: Implement ordered access procedure for CSIS client
the CSIS client now has a ordered access procedure function.
The function implements the procedure as per the CSIP spec,
where it will verify that 1 or more members are unlocked,
and the execture a procedure on each member in ascending
order of rank.

The procedure can be anything (even a non-BT procedure),
and it will be up to the applicaiton to implement
what to do in that.

(cherry picked from commit 7148146348880f862226fc059ec13eac73843f14)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 7148146348880f862226fc059ec13eac73843f14
Change-Id: I64390ade414ee9035d861e7b23ce362e9f7927f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3720861
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>